change some file path rules

+++ src/main/java/ua/net/yason/corpus/sheets/EntryPointRename.java
@@ -17,7 +17,9 @@ import ua.net.yason.corpus.meta.model.TextModel;
  */
 public class EntryPointRename {
 
-    private static String destRootPath = "d:\\Shvedova\\RenamedTexts";
+    private static final String SRC_TEXTS_PATH = "d:\\Shvedova\\Texts";
+    
+    private static String DST_TEXTS_PATH = "d:\\Shvedova\\RenamedTexts";
     
     public static void main(String[] args) throws IOException {
         
@@ -40,11 +42,15 @@ public class EntryPointRename {
             encodedPath = encodedPath.replaceAll("[^\\p{ASCII}]", "");
             encodedPath = charset.decode(charset.encode(encodedPath)).toString();
             encodedPath += ".txt";
+            String firstChar = encodedPath.substring(0, 1).toUpperCase();
+            encodedPath = firstChar + "\\" + encodedPath;
             System.out.println(encodedPath);
-            File srcFile = new File(Options.UKR_CORPUS_TEXTS_ROOT_PATH + "\\" + originPath);
-            File destFile = new File(destRootPath + "\\" + encodedPath);
+            File srcFile = new File(SRC_TEXTS_PATH + "\\" + originPath);
+            File destFile = new File(DST_TEXTS_PATH + "\\" + encodedPath);
             destFile.getParentFile().mkdirs();
+            if (!destFile.exists()){
             FileUtils.copyFile(srcFile, destFile);
+            }
         }
     }
 }
